requirements.txt
aiogram==2.12.1
python-decouple==3.4
pytz